compression/src/main/java/esa/s1pdgs/cpoc/compression/config/ApplicationProperties.java
package esa.s1pdgs.cpoc.compression.config;

import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.context.annotation.Configuration;

@Configuration
@EnableConfigurationProperties
@ConfigurationProperties(prefix = "compression")
public class ApplicationProperties {
	private String command;

	public String getCommand() {
		return command;
	}

	public void setCommand(String command) {
		this.command = command;
	}

}
